+++ packages/core/src/hooks/useLocaleData/examples/DummyComponent.tsx
@@ -1,4 +1,4 @@
-import type { FrontendLocaleData } from "@core";
+import type { FrontendLocaleData } from "@hakit/core";
 
 export function DummyComponentReturn(props: Partial<FrontendLocaleData>) {
   return props;